FIX URL BUG, CREATE DICT FOR LOADING
A bug was found with the URL which was causing the scraper to
scrape the same ebay results page on a loop. Positioning the &_pgn=
field in a different part of the URL, and starting the iteration
from page 2 seems to have fixed this.

Added dictionary creation matching the description to a games 
console, and creating a new entry with title, console and price 
fields.