Add db commands; consolidate ifs into one for loop
Removed multiple loops for each console - consolidated into one loop
through consoles in console list.

Within that loop, title_matcher.check_match() method is called with the game
dict passed in. If a match is found, the game dict is returned and the data is 
loaded into the db. If no match, False is returned and the item is passed over.