Consolidated console searches into one method
Removed check_ps4_data, check_ps3_data etc methods - consolidated
into one check_match method.

This method is very long and convoluted. I am not sure if there is a way
to dynamically refer to an attribute i.e.
for line in self.{console}_data. If there is a way, we can cut this down 
further.