Consolidate check_match method
Replaced method involving multiple if / else statements with
a dictionary with the keys as consoles and the values as class attributes.

The method then passes the attribute as a value to a second method,
which carries out the matching of the game against the relevant
game lists (e.g. self.ps4_data).