Add item ID scrape to scraper
Added id_list to beautifulsoup scraper. Item ID can be used
as an index each time the scraper is run to find the first item
that was scraped the last time the scraper was run i.e. the end
of the range that we want to scrape this time around.

This will prevent the same listing from being double scraped.