Add unique_id_check; change consoles to dict
After adding the id field to the scraper it was found that it was
picking up some duplicate items - possibly because the last 
item on page x is also the first item on page y. Added unique
id list and an additional check within the loop to eliminate
duplicates.

Also changed the console list on line 36 to a dict to standardise
console names at an early stage, so that the console names e.g.
"PS2", "PLAYSTATION 2" are consistent.