Add first_item_index table to db
This table is needed to record the ebay ID of the first result
from each scrape. This will be used to ensure that, when the 
scraper is running at time intervals, the same items will not
be scraped more than once.

Effectively, the last item in the table will act as an index to
tell the scraper to stop when that item is reached.

For example, scrape x will scrape all items from 3100 - 
3000, and record the ebay ID of item 3100. Scrape y
will then scrape from item 3200 - 3101, matching the
ebay ID of 3100 to the last line in the table, telling it
to end the range there.